Update DOCKER_HOST logic
Once DOCKER_HOST is determined, inject it into the os environment directly.
This is currently the only env var that needs to be placed right into the
environment because of chicken-and-egg issues for getting the environment from
a remote docker host.

In order to have the DOCKER_HOST variable when running other commands, it has
to be specially injected into the shell environment, or actions will not take
place on the proper host.