Implement a rendered_data() method
THIS IS A HACK.

The way the env module processes environment is spaghetti.  There are a few things it's doing:

- collecting user-provided variables
- injecting cf-generated variables
- interpolating the config values
- including additional sources of variables

All of these steps are intermixed and the method naming is confusing, and now there's yet another hack to get things appearing to work in a sane way.

The `rendered_data` property will fully-render out the environment and then re-parse it into a dictionary.  Meh.

But it works.  :(